refactor: Refactor and code fixes
This change refactor and also fix the code to support Python 3.7.
In Python 3.7, the `typing` module does not have `Self` class and only supported by
Python 3.11 and later.

To fix this error, we just simply check whether the Python version that runs
this package is 3.11+ or not, if not explicitly declare a `NoneType` global variable with
name the same as `Self` to prevent error.