Introduce a new method in `PomParser` class
The method is called `get`, this method can finds the element tag
from the given key, the key itself can be specified using a string
literal (separate each element using dot '.') or a list of tag names.
The result could be a None, this means that element are undefined
or that users has specified wrong element tree path.

Additionally, we also refactored all methods so they could safely
returns a None value if the desired element are undefined or
cannot be found.